#62 Move decorators to utils module
