+++ azely/object.py
@@ -10,8 +10,8 @@ from typing import Optional
 from astropy.coordinates import Longitude, Latitude, SkyCoord, get_body
 from astropy.time import Time as ObsTime
 from astropy.utils.data import conf
-from .cache import PathLike, cache
 from .consts import AZELY_OBJECTS, FRAME, SOLAR_FRAME, SOLAR_OBJECTS, TIMEOUT
+from .utils import PathLike, cache, rename
 
 
 @dataclass
@@ -83,41 +83,45 @@ def get_object(
         return get_object_by_name(
             query,
             frame=frame,
-            name=name,
             timeout=timeout,
+            name=name,
             source=source,
             update=update,
         )
 
 
+@rename
 @cache
 def get_object_solar(
     query: str,
     /,
     *,
-    name: Optional[str],
-    source: PathLike,  # consumed by @cache
-    update: bool,  # consumed by @cache
+    # consumed by decorators
+    name: Optional[str],  # @rename
+    source: PathLike,  # @cache
+    update: bool,  # @cache
 ) -> Object:
     """Get object information in the solar system."""
     return Object(
-        name=name or query,
+        name=query,
         longitude="NA",
         latitude="NA",
         frame=SOLAR_FRAME,
     )
 
 
+@rename
 @cache
 def get_object_by_name(
     query: str,
     /,
     *,
     frame: str,
-    name: Optional[str],
-    source: PathLike,  # consumed by @cache
     timeout: int,
-    update: bool,  # consumed by @cache
+    # consumed by decorators
+    name: Optional[str],  # @rename
+    source: PathLike,  # @cache
+    update: bool,  # @cache
 ) -> Object:
     """Get object information by an object name."""
     with conf.set_temp("remote_timeout", timeout):
@@ -129,7 +133,7 @@ def get_object_by_name(
         )
 
     return Object(
-        name=name or query,
+        name=query,
         longitude=str(response.data.lon),  # type: ignore
         latitude=str(response.data.lat),  # type: ignore
         frame=frame,
